+++ app/src/main/java/com/henryhiles/qweather/domain/remote/HourlyWeatherDataDto.kt
@@ -6,13 +6,13 @@ data class HourlyWeatherDataDto(
     @field:Json(name = "time")
     val time: List<String>,
     @field:Json(name = "temperature_2m")
-    val temperature: List<Double>,
+    val temperature: List<Float>,
     @field:Json(name = "apparent_temperature")
-    val apparentTemperature: List<Double>,
+    val apparentTemperature: List<Float>,
     @field:Json(name = "weathercode")
     val weatherCode: List<Int>,
     @field:Json(name = "precipitation_probability")
     val precipitationProbability: List<Int>,
     @field:Json(name = "windspeed_10m")
-    val windSpeed: List<Double>,
+    val windSpeed: List<Float>,
 )
